# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kocotree Skills API
  version: 0.5.0
  description: Kocotree Skills 客户端、模拟接口与未来真实后端共同遵守的契约。
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/users/me/skills:
    get:
      tags: [Users]
      summary: 获取我的在线 Skill
      operationId: listMySkills
      security:
        - BearerAuth: []
      parameters:
        - name: relation
          in: query
          required: true
          schema:
            type: string
            enum: [OWNED, COLLABORATED]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Skill 分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillPageResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/users/me/skills/publish-target:
    get:
      tags: [Users]
      summary: 按 Skill 名称解析当前用户可发布的云端目标
      operationId: resolvePublishTarget
      security:
        - BearerAuth: []
      parameters:
        - name: skillName
          in: query
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
            maxLength: 64
      responses:
        '200':
          description: 云端发布目标解析结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTargetResolutionResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/skills/{skillId}:
    get:
      tags: [Skills]
      summary: 获取 Skill 详情
      description: 历史归档数据仅作只读兼容；当前产品不再创建归档状态。
      operationId: getSkill
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      responses:
        '200':
          description: Skill 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetailResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Skills]
      summary: 修改平台展示信息
      operationId: updateSkillMetadata
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSkillRequest'
      responses:
        '200':
          description: Skill 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetailResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Skills]
      summary: 永久删除当前用户拥有的 Skill
      operationId: deleteSkill
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SkillId'
      responses:
        '200':
          description: Skill 已永久删除
          content:
            application/json:
              schema:
                type: object
                required: [code, data, msg]
                properties:
                  code:
                    type: integer
                    const: 200
                  data:
                    type: object
                    required: [id, deletedObjectCount, objectCount, ossCleaned]
                    properties:
                      id:
                        type: string
                      deletedObjectCount:
                        type: integer
                      objectCount:
                        type: integer
                      ossCleaned:
                        type: boolean
                  msg:
                    type: string
                    const: success
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SkillId:
      name: skillId
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ErrorResponse:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, data, msg]
      properties:
        code:
          $ref: '#/components/schemas/ErrorHttpStatus'
        data:
          $ref: '#/components/schemas/ErrorData'
        msg:
          description: 用户可读的错误信息。
          type: string
    ErrorHttpStatus:
      description: 与 HTTP 响应状态码一致。
      type: integer
      enum: [400, 401, 403, 404, 409, 413, 415, 422, 500]
    ErrorData:
      description: 业务错误码与结构化错误详情。
      type: object
      required: [errorCode]
      properties:
        errorCode:
          $ref: '#/components/schemas/BusinessErrorCode'
      additionalProperties: true
    BusinessErrorCode:
      type: string
      enum:
        - CONTENT_UNCHANGED
        - DISPLAY_NAME_CONFIRMATION_REQUIRED
        - DUPLICATE_SKILL_NAME
        - FILE_NOT_FOUND
        - FILE_PREVIEW_UNAVAILABLE
        - FORBIDDEN
        - INSTALLATION_UNAVAILABLE
        - INVALID_REQUEST
        - INVALID_SEMVER
        - INVALID_SKILL_PACKAGE
        - OWNER_REQUIRED
        - PACKAGE_HASH_MISMATCH
        - PACKAGE_TOO_LARGE
        - SKILL_NAME_MISMATCH
        - SKILL_NOT_FOUND
        - SKILL_UNAVAILABLE
        - UNAUTHENTICATED
        - USER_DISABLED
        - VERSION_ALREADY_EXISTS
        - VERSION_CONFLICT
        - VERSION_NOT_FOUND
        - VERSION_NOT_GREATER
    ApiOkMeta:
      type: object
      required: [code, msg]
      properties:
        code:
          description: 与 HTTP 成功状态码一致。
          type: integer
          const: 200
        msg:
          description: 成功响应的固定信息。
          type: string
          const: success
          example: success
...
    User:
      type: object
      required: [id, name, avatarUrl, departmentPath, status, role, syncedAt]
      properties:
        id:
          type: string
        name:
          type: string
        avatarUrl:
          type: [string, 'null']
          format: uri
        departmentPath:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        role:
          type: string
          enum: [USER, ADMIN]
        syncedAt:
          type: string
          format: date-time
    Tag:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 1
          maxLength: 50
    DerivedSource:
      type: object
      required: [skillId, skillName, versionId, version, status, linkable]
      properties:
        skillId:
          type: string
        skillName:
          type: string
        versionId:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [ACTIVE, ARCHIVED, NAME_CONFLICT]
        linkable:
          type: boolean
    SkillVersion:
      type: object
      required:
        - id
        - skillId
        - version
        - status
        - skillName
        - skillDescription
        - changelog
        - baseVersionId
        - packageSize
        - packageSha256
        - contentHash
        - uploadedBy
        - publishedAt
        - withdrawnBy
        - withdrawnAt
        - withdrawalReason
      properties:
        id:
          type: string
        skillId:
          type: string
        version:
          type: string
          pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)(?:-[0-9A-Za-z.-]+)?(?:\+[0-9A-Za-z.-]+)?$'
        status:
          type: string
          enum: [PUBLISHED, WITHDRAWN]
        skillName:
          type: string
          maxLength: 64
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        skillDescription:
          type: string
          minLength: 1
        changelog:
          type: string
          minLength: 1
          maxLength: 2000
        baseVersionId:
          type: [string, 'null']
        packageSize:
          type: integer
          minimum: 1
          maximum: 52428800
        packageSha256:
          type: string
        contentHash:
          type: string
        uploadedBy:
          $ref: '#/components/schemas/User'
        publishedAt:
          type: string
          format: date-time
        withdrawnBy:
          oneOf:
            - $ref: '#/components/schemas/User'
            - type: 'null'
        withdrawnAt:
          type: [string, 'null']
          format: date-time
        withdrawalReason:
          type: [string, 'null']
...
    SkillSummary:
      type: object
      required:
        - id
        - skillName
        - displayName
        - skillDescription
        - displayDescription
        - status
        - owner
        - tags
        - currentVersion
        - installCount
        - derivedFrom
        - updatedBy
        - archivedAt
        - archiveReason
        - nameConflictReason
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        skillName:
          type: string
          maxLength: 64
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        skillDescription:
          type: string
          minLength: 1
        displayDescription:
          type: string
          minLength: 1
          maxLength: 1000
        status:
          type: string
          enum: [ACTIVE, ARCHIVED, NAME_CONFLICT]
        owner:
          $ref: '#/components/schemas/User'
        tags:
          type: array
          minItems: 1
          maxItems: 5
          items:
            $ref: '#/components/schemas/Tag'
        currentVersion:
          $ref: '#/components/schemas/SkillVersion'
        installCount:
          type: integer
          minimum: 0
        derivedFrom:
          oneOf:
            - $ref: '#/components/schemas/DerivedSource'
            - type: 'null'
        updatedBy:
          $ref: '#/components/schemas/User'
        archivedAt:
          type: [string, 'null']
          format: date-time
        archiveReason:
          type: [string, 'null']
        nameConflictReason:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SkillDetail:
      allOf:
        - $ref: '#/components/schemas/SkillSummary'
        - type: object
          required: [collaborators, derivedChain]
          properties:
            collaborators:
              type: array
              items:
                $ref: '#/components/schemas/User'
            derivedChain:
              type: array
              items:
                $ref: '#/components/schemas/DerivedSource'
...
    UpdateSkillRequest:
      type: object
      minProperties: 1
      anyOf:
        - required: [displayName]
        - required: [displayDescription]
        - required: [tagIds]
        - required: [newTagNames]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        displayDescription:
          type: string
          minLength: 1
          maxLength: 1000
        tagIds:
          type: array
          minItems: 0
          maxItems: 5
          items:
            type: string
        newTagNames:
          type: array
          minItems: 0
          maxItems: 5
          items:
            type: string
            minLength: 1
            maxLength: 50
        confirmDuplicateDisplayName:
          type: boolean
          default: false
...
    PublishTargetResolution:
      type: object
      required: [state, skill]
      properties:
        state:
          type: string
          enum: [OWNED, NOT_FOUND, TAKEN_BY_OTHER, UNAVAILABLE]
        skill:
          oneOf:
            - $ref: '#/components/schemas/SkillDetail'
            - type: 'null'
...
    SkillPage:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SkillSummary'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    SkillPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiOkMeta'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/SkillPage'
...
    SkillDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiOkMeta'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/SkillDetail'
...
    PublishTargetResolutionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiOkMeta'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/PublishTargetResolution'
...